Fixed issue
The issue was with grabbing local materialse.css and materialize.js files and the web server was seeing them as insecure when not grabbed from the cache. The initial load now happens through a CDN and fixes the issue. Shouldn;'t affect the speed of the site as it only happens the first time the site is loaded and then they are cached