Timer runs in the background
The time r now runs in the background using html5 web workers. The timer functionality is passed off to the web worker and then the changing of the DOM is managed by the script that started the web worker. When pausing and/or resetting the timer, the values are changed on the DOM by app.js, and the values on the timer itself are changed by a postMessage() call to the background_process.js